fixed, LOOK AT DESCRIPTION IN CASE OF BUG, I GOT IDEAS
the bug was actually not cause of the angles, it was because the 0th
step looked at the other wall due to a large offs_x / offs_y.
I just removed the 0th step. I know this will come back once ive got
movement and can stant close to walls, however, i already got a couple
ideas for fixes in my head for that:
1) dont allow the player to get that close (if that doesnt look too
wrong)
2) make a special check for the 0th step, where basically just the a 3x3
grid around the ray position is checked (or rather a 2x2 with the origin
in one corner)
3) figure something else out, ure probably smarter than me, future me :3