README || a bunch of changes that dont matter for the user
Part of this software, more specifically the part that display.c was/is used for here, has now been moved to a separate repo, https://github.com/IOKG04/MaDiEn. Once I am done adding some features to that, I will likely integrade it here, but for now expect me to be there mostly, as i add new features to it.
After that tho, so i dont forget what i have planned for this, heres my todo list:
- Add wall collision
- Make an algorithm that determines a goal tile
  - Make sure the goal is accessible (pathfinding maybe)
- Add a "u win" screen (no fail screen cause u cant fail)
- Add a menu and saves
- JUST MAKE IT BETTER
-----
as well as the ability to click 7 and 9 to turn 90 degrees in a specific direction.
in case you dont get those numbers, GO BUY A NUMPAD!!!!